add validator for state list
